backend/requirements.txt
Django==5.2.4
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
django-cors-headers==4.3.1
python-decouple==3.8
pdfplumber==0.10.3
Pillow==10.1.0
reportlab==4.0.7 